+++ assets/asset.go
@@ -14,9 +14,13 @@ type isAsset interface {
 	GetAssetClass() common.AssetClass
 	// IsValid is the order valid / validated with the target chain?
 	IsValid() bool
-	// Validate this is used to validate that the asset
-	// exist on the target chain
+	// Validate this is used to check if the assets
+	// are present on the target chain
 	Validate() error
+	// SetValidNonValidator will set an asset as valid
+	// without running actual validation, this is used in the
+	// context of a non-validator node.
+	SetValidNonValidator()
 	String() string
 }