+++ cmd/vega/genesis/update.go
@@ -26,7 +26,7 @@ type updateCmd struct {
 
 	DryRun  bool   `long:"dry-run" description:"Display the genesis file without writing it"`
 	Network string `short:"n" long:"network" choice:"mainnet" choice:"testnet"`
-	TmRoot  string `short:"t" long:"tm-root" description:"The root path of tendermint"`
+	TmHome  string `short:"t" long:"tm-home" description:"The home path of tendermint"`
 }
 
 func (opts *updateCmd) Execute(_ []string) error {
@@ -53,7 +53,7 @@ func (opts *updateCmd) Execute(_ []string) error {
 
 	// genesis file
 	tmConfig := tmconfig.DefaultConfig()
-	tmConfig.SetRoot(os.ExpandEnv(opts.TmRoot))
+	tmConfig.SetRoot(os.ExpandEnv(opts.TmHome))
 
 	pubKey, err := loadTendermintPrivateValidatorKey(tmConfig)
 	if err != nil {