+++ netparams/defaults.go
@@ -123,8 +123,9 @@ func defaultNetParams() map[string]value {
 		NetworkCheckpointMarketFreezeDate:              NewTime().Mutable(true).MustUpdate("never"),
 		NetworkCheckpointNetworkEOLDate:                NewTime().Mutable(true).MustUpdate("never"),
 		NetworkCheckpointTimeElapsedBetweenCheckpoints: NewDuration(DurationGT(0 * time.Second)).Mutable(true).MustUpdate("1m"),
-		// take a snapshot every N blocks, default to every block for now (debug)
-		SnapshotIntervalLength: NewInt(IntGTE(0)).Mutable(true).MustUpdate("1"),
+		// take a snapshot every 1000 blocks, ~20 minutes
+		// if we assume a block time of anything between 1 to 2 seconds
+		SnapshotIntervalLength: NewInt(IntGTE(0)).Mutable(true).MustUpdate("1000"),
 
 		FloatingPointUpdatesDuration: NewDuration().Mutable(true).MustUpdate("5m"),
 	}